Rename: zzal 폴더 내 신고 관련 api 요청 데이터 전송 객체를 report폴더로 이동
