+++ src/apis/report.ts
@@ -1,4 +1,4 @@
-import { PostReportZzalRequeust } from "@/types/zzal.dto";
+import { PostReportZzalRequeust } from "@/types/report.dto";
 import http from "./core";
 
 export const postReportZzal = ({ imageId }: PostReportZzalRequeust) => {
