+++ src/types/report.dto.ts
@@ -1,4 +1,6 @@
-import { Report, ReportDetail } from "./report";
+import { Report, ReportDetail, ReportZzal } from "./report";
 
 export type GetReportResponse = Report[];
 export type GetReportDetailResponse = ReportDetail[];
+
+export type PostReportZzalRequeust = ReportZzal;
