--- src/types/zzal.dto.ts
@@ -1,3 +1,0 @@
-import { ReportZzal } from "./zzal";
-
-export type PostReportZzalRequeust = ReportZzal;
