--- src/types/zzal.ts
@@ -1,3 +1,0 @@
-export interface ReportZzal {
-  imageId: number;
-}
